// src/components/types.ts
export type Orientation = "vertical" | "horizontal";

export type AxisType = "value" | "category";

export type HorizontalPosition = "left" | "right";
export type VerticalPosition = "top" | "bottom";

export type LegendType = "plain" | "scroll";

export type Icon =
  | "circle"
  | "rect"
  | "roundRect"
  | "triangle"
  | "diamond"
  | "pin"
  | "arrow"
  | "none";

export type TooltipTrigger = "item" | "axis" | "none";

export type TooltipAxisPointerType = "line" | "shadow" | "none" | "cross";